pic de glace sur soldat (should be good à tester)

+++ Assets/Scripts/Enemies/BaseComportement/EnemyHealth.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour
 {
-    [SerializeField] private int hp;
+    [SerializeField] private float hp;
     [SerializeField] string elevatorToUnlock;
     [SerializeField] bool needUnlock;
     [SerializeField] int indexIceBar;
@@ -36,7 +36,7 @@ public class EnemyHealth : MonoBehaviour
         GameManager.instance.AddToList(indexIceBar, gameObject);
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (!isBlocking)
         {
